Conflicts:
	src/main/webapp/assets/js/ReviewList.js
	src/main/webapp/index.html

+++ src/main/java/com/deli/deliverypp/controller/ProductController.java
@@ -2,15 +2,20 @@ package com.deli.deliverypp.controller;
 
 import com.deli.deliverypp.DB.DeliUser;
 import com.deli.deliverypp.model.Product;
+import com.deli.deliverypp.model.Store;
 import com.deli.deliverypp.service.StoreService;
 import com.deli.deliverypp.util.ControlUtil;
+import com.deli.deliverypp.util.MessageGenerator;
 import com.deli.deliverypp.util.annotaions.ProtectedResource;
 import com.deli.deliverypp.util.annotaions.RequiredModel;
+import com.deli.deliverypp.util.exp.AuthorityChecker;
 
 import javax.servlet.*;
 import javax.servlet.http.*;
 import javax.servlet.annotation.*;
 import java.io.IOException;
+
+import static com.deli.deliverypp.util.JSONUtil.getMapper;
 
 @WebServlet(name = "ProductController", value = "/products/*")
 public class ProductController extends HttpServlet {
@@ -62,7 +67,19 @@ public class ProductController extends HttpServlet {
     // UPDATE product
     @Override
     protected void doPut(HttpServletRequest req, HttpServletResponse resp){
-        ControlUtil.responseMsg(resp, service.updateProduct(ControlUtil.getJson(req)));
+        String json = ControlUtil.getJson(req);
+        try {
+            Product product = getMapper().readValue(json, Product.class);
+            if (AuthorityChecker.checkUserEmail(req, Store.class, "storeId", product.getStoreId())) {
+                ControlUtil.responseMsg(resp, service.updateProduct(json));
+            }else {
+                ControlUtil.sendUnAuthorizeMsg(resp);
+            }
+
+        } catch (Exception e) {
+            ControlUtil.sendResponseData(resp, MessageGenerator.makeErrorMsg("형식이 잘못되었습니다", "type_error"));
+            e.printStackTrace();
+        }
     }
 
     @RequiredModel(target = Product.class)
@@ -70,8 +87,18 @@ public class ProductController extends HttpServlet {
     // DELETE Product
     @Override
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp){
+        String pid = ControlUtil.getRequestUri(req);
+        Product product = service.getProductById(pid);
+        String sid = "";
+        if (product != null) {
+            sid = product.getStoreId();
+        }
+        if (AuthorityChecker.checkUserEmail(req, Store.class, "storeId", sid)){
+            ControlUtil.responseMsg(resp, service.deleteProduct(ControlUtil.getRequestUri(req, 1)));
+        }else {
+            ControlUtil.sendUnAuthorizeMsg(resp);
+        }
         System.out.println(ControlUtil.getRequestUri(req, 1));
-        ControlUtil.responseMsg(resp, service.deleteProduct(ControlUtil.getRequestUri(req, 1)));
     }
