+++ src/main/java/com/deli/deliverypp/controller/review/ReviewController.java
@@ -1,14 +1,17 @@
 package com.deli.deliverypp.controller.review;
 
 import com.deli.deliverypp.DB.DeliUser;
+import com.deli.deliverypp.auth.AuthProvider;
 import com.deli.deliverypp.model.ResponseMessage;
 import com.deli.deliverypp.model.Review;
 import com.deli.deliverypp.service.ReviewService;
 import com.deli.deliverypp.util.ControlUtil;
+import com.deli.deliverypp.util.MessageGenerator;
 import com.deli.deliverypp.util.annotaions.PathParam;
 import com.deli.deliverypp.util.annotaions.ProtectedResource;
 import com.deli.deliverypp.util.annotaions.RequiredModel;
 import com.deli.deliverypp.util.annotaions.RequiredParam;
+import com.deli.deliverypp.util.exp.AuthorityChecker;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -24,6 +27,7 @@ public class ReviewController extends HttpServlet {
 
     private static final ReviewService service = new ReviewService();
     private static final Logger log = LogManager.getLogger(ReviewController.class);
+    private static final AuthProvider provider = new AuthProvider();
 
     // option query string
     @Override
@@ -36,35 +40,49 @@ public class ReviewController extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        ControlUtil.sendResponseData(response, service.insertNewReview(ControlUtil.getJson(request), true));
+        String json = ControlUtil.getJson(request);
+        DeliUser user = provider.getUserFromHeader(request);
+
+        log.info(json);
+        log.info(user);
+        if (user != null) {
+            ControlUtil.sendResponseData(response, service.insertNewReview(json, user.getUserEmail()));
+        }else {
+            ControlUtil.sendResponseData(response, service.insertNewReview(json, null));
+        }
     }
 
 
 
     @RequiredModel(target = Review.class)
     @ProtectedResource(uri = "/review", id = true, method = "put")
+    @RequiredParam(value = "json")
     @Override
     protected void doPut(HttpServletRequest req, HttpServletResponse resp) {
-        ControlUtil.sendResponseData(resp, service.updateReview(ControlUtil.getJson(req)));
+        String json = ControlUtil.getJson(req);
+        if (AuthorityChecker.checkUserEmailFromJson(req,Review.class,"reviewId", json)){
+            ControlUtil.sendResponseData(resp, service.updateReview(json));
+        }else {
+            ControlUtil.sendResponseData(resp, MessageGenerator.makeErrorMsg("you don't have authority to access this resources \r\n 본인인증실패", "authority_error"));
+            log.warn("failed to pass test");
+        }
     }
 
 
 
+    // TODO REFACTORING !!!!
     @ProtectedResource(uri = "/review", id = true, method = "delete")
     @RequiredParam(value = "/review/{reviewId}")
     @Override
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp){
-        System.out.println(ControlUtil.getRequestUri(req, 2));
-        log.info(req.getRequestURI());
-        ControlUtil.responseMsg(resp, service.deleteReview(ControlUtil.getRequestUri(req)));
-    }
-
-
-
-//    @RequiredParam(value = "/review/{reviewId}")
-//    @ProtectedResource(uri = "/review", id = true, method = "delete")
-    public void delete(@PathParam(param = "reviewId") String reviewId) {
-        System.out.println(reviewId);
+        String reviewId = ControlUtil.getRequestUri(req);
+        if (AuthorityChecker.checkUserEmail(req,Review.class,"reviewId", reviewId)){
+            log.warn("passed annotaions test");
+            ControlUtil.responseMsg(resp, service.deleteReview(ControlUtil.getRequestUri(req)));
+        }else{
+            ControlUtil.sendResponseData(resp, MessageGenerator.makeErrorMsg("you don't have authority to access this resources \r\n 본인인증실패", "authority_error"));
+            log.warn("failed to pass test");
+        }
     }
 
 
@@ -73,7 +91,7 @@ public class ReviewController extends HttpServlet {
         String target = ControlUtil.getRequestUri(request);
         String value = ControlUtil.getRequestUri(request, 2);
         ResponseMessage<?> msg = null;
-        System.out.println(value);
+        log.debug(value);
         switch (target) {
             case "user":
                 msg = service.getReviewsByWriter(value);
